// src/types/StoreTypes.tsx
export interface peopleInputState {
    id: number;
    name: string;
    birth: string | null;
    gender: string | null;
    calendarType: string | null;
    registration: string | null;
    maritalStatus: string | null;
    position: string | null;
    department: string | null;
    department2: string | null;
    phone1: string | null;
    phone2: string | null;
    spouse: string | null;
    address1: string | null;
    address2: string | null;
    note: string | null;
    
}